Fix windows style V2

+++ valheim/valheim_bepinex_hexium/install_script.sh
@@ -206,13 +206,12 @@ if [ ! -z "$V_MODPACK_URL" ]; then
         fi
 
         # Fix Windows-style backslashes in extracted paths
-        find "$DEPENDENCY_TEMP_DIR" -depth -name '*\*' -exec bash -c '
-            for FILE; do
-                NEW_FILE="${FILE//\\//}"
-                mkdir -p "$(dirname "$NEW_FILE")"
-                mv "$FILE" "$NEW_FILE"
-            done
-        ' bash {} +
+        for FILE in "$DEPENDENCY_TEMP_DIR"/*\\*; do
+            [ -e "$FILE" ] || continue
+            NEW_FILE="${FILE//\\//}"
+            mkdir -p "$(dirname "$NEW_FILE")"
+            mv "$FILE" "$NEW_FILE"
+        done
 
         # Check if the extracted directory contains BepInEx folder or individual plugin folders
         if [ -d "$DEPENDENCY_TEMP_DIR/BepInEx" ]; then
